+++ R/rc_helpers.R
@@ -359,7 +359,7 @@ rc_check_inputs <- function(soil_properties,
   if(!is.null(rothc_rotation)){
     checkmate::assert_data_table(rothc_rotation, null.ok = TRUE, min.rows = 1)
 
-    req <- c("B_LU_START", "B_LU_END", "B_LU","B_LU_HC","B_C_OF_INPUT")
+    req <- c("B_LU_START", "B_LU_END", "B_LU_HC","B_C_OF_INPUT")
     checkmate::assert_names(colnames(rothc_rotation), must.include = req)
     
     checkmate::assert_numeric(rothc_rotation$B_LU_HC, lower = rc_minval('B_LU_HC'), upper = rc_maxval('B_LU_HC'), any.missing = FALSE)
@@ -394,7 +394,20 @@ rc_check_inputs <- function(soil_properties,
 #' @param dt (data table) Contains the columns A_CLAY_MI (clay content \%) and at least one of A_SOM_LOI (organic matter content, \%) and A_C_OF (organic carbon content, g C/kg)
 #'
 #' @returns
-#' Data table with the dry soil bulk density (g/cm3)
+#' Data table with provided data and the calculated dry soil bulk density (g/cm3)
+#' 
+#' @examples
+#' ## calculate the dry soil bulk density (g/cm3) using organic matter content (\%)
+#' dt_om <- data.table::data.table(A_CLAY_MI = 12, # data table with legal soil information
+#' A_SOM_LOI = 3)
+#' 
+#' rc_calculate_bd(dt_om)
+#' 
+#' ## calculate the dry soil bulk density using organic carbon content (g C/kg)
+#' dt_c <- data.table::data.table(A_CLAY_MI = 12, # data table with legal soil information
+#' A_C_OF = 80)
+#' 
+#' rc_calculate_bd(dt_c)
 #' 
 #' @export
 rc_calculate_bd <- function(dt){
@@ -427,6 +440,10 @@ rc_calculate_bd <- function(dt){
   
   # clay dependent density
   dt[, A_DENSITY_SA := cf * dens.clay + (1-cf) * dens.sand]
+  
+  # select relevant columns
+  dt <- dt[, .SD, .SDcols =!names(dt) %in% c("dens.sand", "dens.clay", "cf")]
+  
   return(dt)
 }
 
@@ -435,7 +452,7 @@ rc_calculate_bd <- function(dt){
 #' @param dt (data table) Table with crop rotation details and crop management actions that have been taken. For more information see details.
 #'
 #' @returns
-#' Data table with total C input from crops
+#' Data table with provided input together with calculated C input from crops
 #' 
 #' @details
 #' Crop data table
@@ -446,6 +463,17 @@ rc_calculate_bd <- function(dt){
 #' * B_LU_RS_FR (numeric), Root-to-shoot ratio of the crop
 #' * M_CROPRESIDUE (logical), indicator of whether crop residue is incorporated into the soil
 #' 
+#' @examples
+#' ## calculate crop C inputs with valid input values
+#'  dt_crop <- data.table::data.table(B_LU_YIELD = 30000,
+#'  B_LU_HI = 0.6,
+#'  B_LU_HI_RES = 0.5,
+#'  B_LU_RS_FR = 1,
+#'  M_CROPRESIDUE = TRUE
+#'  )
+#'  
+#'  rc_calculate_B_C_OF(dt_crop)
+#'  
 #' @export
 
 rc_calculate_B_C_OF <- function(dt){
@@ -480,8 +508,8 @@ return(dt.crop)
 #'
 #' @param crops (data.table) Input crop table for the RothC model. See details for further information
 #' @param simyears (numeric) number of simulation years of the RothC model
-#' @param start_date (character, formatted YYYY-MM-DD) Date in which simulation starts
-#' @param end_date (character, formatted YYYY-MM-DD) Required if simyears is not supplied
+#' @param start_date (character/date, formatted YYYY-MM-DD) Date in which simulation starts
+#' @param end_date (character/date, formatted YYYY-MM-DD) Required if simyears is not supplied
 #'
 #' @returns
 #' An extended crop input file to be used in the rotsee package
@@ -491,11 +519,26 @@ return(dt.crop)
 #' Includes the columns: 
 #' * B_LU_START (start of crop rotation),
 #' * B_LU_END (end of crop rotation),
-#' * B_LU (a crop id), 
+#' * B_LU (a crop id) optional, 
 #' * B_LU_NAME (a crop name, optional),
 #' * B_LU_HC, the humification coefficient of crop organic matter (-). When not supplied, default RothC value will be used
 #' * B_C_OF_INPUT, the organic carbon input on field level (kg C/ha)
 #' 
+#' @examples
+#' ## examples to extend valid crop dataset
+#' # create valid crop set
+#' crops <- data.table::data.table(
+#' B_LU_START = c("2020-01-01", "2020-06-01"),
+#' B_LU_END = c("2020-03-31", "2020-08-31"),
+#' B_LU_HC = c(0.5, 0.3),
+#' B_C_OF_INPUT = c(100, 200)
+#' )
+#' 
+#' # using start_date and end_date
+#' rc_extend_crops(crops = crops, start_date = "2020-01-01", end_date = "2022-12-31")
+#' 
+#' # using start_date and simyears
+#' rc_extend_crops(crops = crops, start_date = "2020-01-01", simyears = 3)
 #' 
 #' @export
 #'
@@ -510,7 +553,7 @@ rc_extend_crops <- function(crops,start_date, end_date = NULL, simyears = NULL){
   crops <- as.data.table(crops)
   setnames(crops,toupper(colnames(crops)))
   
-  req <- c("B_LU_START", "B_LU_END", "B_LU", "B_LU_HC", "B_C_OF_INPUT")
+  req <- c("B_LU_START", "B_LU_END", "B_LU_HC", "B_C_OF_INPUT")
   checkmate::assert_names(colnames(crops), must.include = req)
   if ("B_LU_NAME" %in% names(crops)) checkmate::assert_character(crops$B_LU_NAME, any.missing = FALSE)
   checkmate::assert_numeric(crops$B_LU_HC, lower = rc_minval('B_LU_HC'), upper = rc_maxval('B_LU_HC'), any.missing = F)
@@ -598,11 +641,27 @@ rc_extend_crops <- function(crops,start_date, end_date = NULL, simyears = NULL){
 #' Includes the columns:
 #' * P_ID (character), ID of the soil amendment product
 #' * P_NAME (character), name of the soil amendment product, optional
-#' * P_C_OF_INPUT (numeric), the organic carbon input from soil amendment product on a field level (kg C/ha)
+#' * B_C_OF_INPUT (numeric), the organic carbon input from soil amendment product on a field level (kg C/ha)
 #' * P_DOSE (numeric), applied dose of soil amendment product (kg/ha), required if P_C_OF_INPUT is not supplied
 #' * P_C_OF (numeric), organic carbon content of the soil amendment product (g C/kg), required if P_C_OF_INPUT is not supplied
 #' * P_HC (numeric), the humification coefficient of the soil amendment product (fraction)
 #' * P_DATE_FERTILIZATION (date), date of fertilizer application (formatted YYYY-MM-DD)
+#' 
+#' #' @examples
+#' ## examples to extend valid amendment dataset
+#' # create valid amendment set
+#' amendments <- data.table::data.table(
+#' P_HC = c(0.5, 0.3),
+#' P_DATE_FERTILIZATION = as.Date(c("2020-01-01", "2020-06-01")),
+#' P_DOSE = c(100, 200),
+#' P_C_OF = c(150,100)
+#' )
+#' 
+#' # using start_date and end_date
+#' rc_extend_amendments(amendments = amendments, start_date = as.Date("2020-01-01"), end_date = as.Date("2022-12-31"))
+#' 
+#' # using start_date and simyears
+#' rc_extend_amendments(amendments = amendments, start_date = as.Date("2020-01-01"), simyears = 3)
 #' 
 #' @export
 #'
@@ -843,6 +902,7 @@ if(!is.null(event)){
   
   # Save plots to files
   if (!dir.exists(save_dir)) dir.create(save_dir, recursive = TRUE, showWarnings = FALSE)
+  message("saving to: ", save_dir)
   ggplot2::ggsave(file.path(save_dir, "carbon_pools_linear.png"), plot = out_cont, width = 10, height = 6)
   ggplot2::ggsave(file.path(save_dir, "carbon_pools_change.png"), plot = out_change_cont, width = 10, height = 6)